+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.3 - 2026-05-29
 
 - fix: support Frostmoon rings on virtual fingers
+- fix: ignore zero-magnitude enchantments in strength scaling
 
 ## 0.3.2 - 2026-05-29